fix: add min hw specs for bootstrap node
I ran into an issue during installation where the `ako` pod was never
scheduled due to the Kind nodes having insufficient CPU.

Unfortunately, none of the `Deployment`s defined for the providers that
come with TKG have any `requests` or `limits` defined, so it is not
possible to calculate the upper-bound of resources required to stand up
a fully-operational bootstrap cluster.

I think 2 vCPUs should be enough, but this might need to increase in the
future.