Add TKGm Azure Hybrid RA addendum and deploy guide
A TKO on Azure reference architecture addendum to add Azure-Hybrid
concepts. Notably, facilitates Private DNS for on-prem resolution as
well as ensuring architecture and security for a "private IP" cluster.
Proxy definitions are allowed for within the solution and a structured
cluster configuration YAML file is provided. USER MUST PROVIDE DETAILS
FOR VARIABLES AND SCRIPTS AS REPRESENTED WITHIN THE ATTACHED DOCS.

Signed-off-by: Olaf Gradin <ogradin@vmware.com>